psi_gen: same pmt pid had continue cc

+++ include/section.h
@@ -6,7 +6,7 @@
 
 int section_to_ts_length(int sec_len);
 int section_to_ts(uint8_t *sec_data, int sec_len, uint8_t *ts_buf,
-					 uint16_t pid, int *p_cc);
+					 uint16_t pid, uint8_t *p_cc);
 int ts_to_section(uint8_t *ts, uint8_t *sec);
